💩 refactor : FeedCardList와 getFeedType으로 리펙토링
